requirements.txt
# django
django
django-rest-framework
django-cors-headers

# for user authentication (password hashing)
bcrypt
